Refactor sendEmail
