+++ src/utils/email/index.ts
@@ -1,1 +1,2 @@
 export * from './emailBuilder';
+export * from './sendEmail';
